agents/requirements.txt
crewai==0.80.0
crewai-tools==0.14.0
langchain==0.3.0
langchain-groq==0.2.0
langchain-core==0.3.0
pydantic==2.10.0
python-dotenv==1.0.0
firebase-admin==6.5.0
google-cloud-firestore==2.19.0
requests==2.32.0
apscheduler==3.10.4
python-telegram-bot==21.6
openai==1.55.0
pillow==11.0.0
boto3==1.35.0
google-api-python-client==2.154.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
